Splitting long message in short messages
